PODS-5758 Fixing more issues
